#42 [fix] 코리 반영 및 iconType 검사 분리
