+++ src/interfaces/event/EventCreateDto.ts
@@ -2,5 +2,5 @@ export interface EventCreateDto {
   eventName: string;
   eventIcon: string;
   date: Date;
-  participant: string[];
+  participants: string[];
 }
