+++ src/services/EventServiceUtils.ts
@@ -1,7 +1,6 @@
 import errorGenerator from '../errors/errorGenerator';
 import Room from '../models/Room';
 import User from '../models/User';
-import { IconTypeArray } from '../modules/IconType';
 import message from '../modules/responseMessage';
 import statusCode from '../modules/statusCode';
 
@@ -27,17 +26,7 @@ const findRoomById = async (roomId: string) => {
   return room;
 };
 
-const isIconType = (iconType: string) => {
-  if (IconTypeArray.indexOf(iconType) == -1) {
-    throw errorGenerator({
-      msg: message.INVALID_ICON_ENUM,
-      statusCode: statusCode.BAD_REQUEST
-    });
-  }
-};
-
 export default {
   findUserById,
-  findRoomById,
-  isIconType
+  findRoomById
 };
